tidying-up files & project structure | add components to landing page
