+++ src/pages/Landing/components/Features.jsx
@@ -1,200 +1,15 @@
 import {
   CheckBadgeIcon,
-  ChevronUpIcon,
   FireIcon,
-  HeartIcon,
   PencilIcon,
   PlusIcon,
   TrashIcon,
 } from "@heroicons/react/24/outline";
-import { cn, getPast7Days, getWeekDayString, nextState } from "../utils";
-import { Link, Navigate } from "react-router-dom";
-import { useAuth } from "../context/AuthContext";
-import { AnimatePresence, motion } from "framer-motion";
-import { IconButton } from "../components/Buttons";
-import { useEffect, useState } from "react";
-import { ENTRY_STATE } from "../constants";
-
-const Landing = () => {
-  const { user } = useAuth();
-
-  if (user) {
-    return <Navigate to={"/habits"} />;
-  }
-
-  return (
-    <main className="scroll-smooth max-w-7xl mx-auto px-4">
-      <Header />
-      <Hero />
-      <Features />
-      <HowItWorks />
-      <GetStarted />
-      <Footer />
-      <ScrollToTop />
-    </main>
-  );
-};
-
-const ScrollToTop = () => {
-  const [isVisible, setIsVisible] = useState(false);
-
-  useEffect(() => {
-    const onScroll = () => {
-      setIsVisible(window.scrollY > 200);
-    };
-
-    window.addEventListener("scroll", onScroll);
-
-    return () => {
-      window.removeEventListener("scroll", onScroll);
-    };
-  }, []);
-
-  const scrollUp = () => {
-    window.scroll({ top: 0, behavior: "smooth" });
-  };
-
-  return (
-    <AnimatePresence>
-      {isVisible && (
-        <motion.button
-          initial={{ opacity: 0, scale: 0, y: 10 }}
-          animate={{
-            opacity: 1,
-            scale: 1,
-            y: 0,
-          }}
-          whileHover={{ scale: 1.05 }}
-          exit={{ opacity: 0, scale: 0, y: 10 }}
-          onClick={scrollUp}
-          className={
-            "fixed border bg-zinc-900 z-10 h-10 w-10 rouded-md grid place-content-center hover:border-white/20 border-white/10 rounded-md bottom-10 right-10 hover:text-white group transition-colors duration-200"
-          }
-        >
-          <ChevronUpIcon className="h-5 w-10 text-zinc-400 transition-colors duration-200 group-hover:text-white group-hover:scale-105" />
-        </motion.button>
-      )}
-    </AnimatePresence>
-  );
-};
-
-const navItems = [
-  {
-    name: "Features",
-    href: "#features",
-  },
-  {
-    name: "How it works",
-    href: "#how-it-works",
-  },
-  {
-    name: "Get started",
-    href: "#get-started",
-  },
-];
-
-const Header = () => {
-  const { user, isLoading } = useAuth();
-
-  return (
-    <header className="flex h-20  items-center">
-      <Link to={"/"} className="font-bold leading-none text-xl text-zinc-100">
-        Doguihabits
-      </Link>
-
-      <ul className="hidden sm:ml-10 sm:flex sm:justify-self-center sm:gap-4">
-        {navItems.map((item) => {
-          return (
-            <li key={item.name}>
-              <button
-                onClick={() => {
-                  document
-                    .querySelector(item.href)
-                    .scrollIntoView({ behavior: "smooth" });
-                }}
-                className="active:text-white leading-none font-medium text-zinc-400 hover:text-white transition-color duration-150"
-                href={item.href}
-              >
-                {item.name}
-              </button>
-            </li>
-          );
-        })}
-      </ul>
-
-      {!isLoading && (
-        <Link
-          className={
-            "bg-emerald-600 font-medium h-10 px-6 grid place-content-center rounded-md text-white ml-auto"
-          }
-          to={user ? "/habits" : "/login"}
-        >
-          {user ? "My Habits" : "Sign Up Now"}
-        </Link>
-      )}
-    </header>
-  );
-};
-
-const Hero = () => {
-  return (
-    <section className="relative min-h-[calc(100vh-_5rem)] py-36 md:py-40 flex flex-col items-center">
-      <div className="min-w-fit text-center z-10">
-        <span className="uppercase tracking-widest block  text-emerald-500">
-          doguihabits
-        </span>
-        <h1 className="text-4xl md:text-5xl font-bold text-white max-w-[18ch] leading-tight">
-          Your Personal Habit Transformation Journey
-        </h1>
-        <p className="md:text-lg text-zinc-400 font-medium mt-6 md:mt-10 max-w-[50ch]">
-          Welcome to doguihabits, the simple yet powerful habit tracker designed
-          to empower your journey towards personal growth and transformation.
-        </p>
-        <div className="flex gap-4 justify-center items-center mt-12">
-          <Link
-            className={
-              "bg-emerald-600 font-medium h-10 px-6 grid place-content-center rounded-md text-white "
-            }
-            to={"/login"}
-          >
-            Sign Up
-          </Link>
-          <Link
-            className={
-              "font-medium h-10 px-6 grid place-content-center rounded-md text-white "
-            }
-            to={"/contact"}
-          >
-            Contact us
-          </Link>
-        </div>
-      </div>
-
-      <div className="absolute bottom-0 overflow-hidden  border-green-600/5 w-full">
-        <motion.img
-          initial={{ opacity: 0, y: 10 }}
-          animate={{ opacity: 1, y: 0 }}
-          transition={{ duration: 0.5, ease: "easeInOut" }}
-          src={"/herov2.svg"}
-          className="hidden md:block mx-auto select-none pointer-events-none"
-        />
-        <motion.img
-          initial={{ opacity: 0, y: 10 }}
-          animate={{ opacity: 1, y: 0 }}
-          transition={{ duration: 0.5, ease: "easeInOut" }}
-          src={"/herov2mobile.svg"}
-          className="block md:hidden mx-auto select-none pointer-events-none"
-        />
-      </div>
-
-      <div>
-        <div className="h-4 w-32 block absolute -bottom-2 left-0 bg-gradient-to-r from-zinc-900 z-10"></div>
-        <div className="h-4 w-32 block absolute -bottom-2 right-0 bg-gradient-to-l from-zinc-900 z-10 "></div>
-        <span className="border-t absolute left-0  right-0 w-full bottom-0 border-white/10 block" />
-      </div>
-    </section>
-  );
-};
+import { cn, getPast7Days, getWeekDayString, nextState } from "../../../utils";
+import { motion } from "framer-motion";
+import { ENTRY_STATE } from "../../../constants";
+import { IconButton } from "../../../components/Buttons";
+import { useState } from "react";
 
 const Features = () => {
   return (
@@ -441,144 +256,4 @@ function WeekView({ habitTitle, className }) {
   );
 }
 
-const Layout = () => {
-  return <div></div>;
-};
-
-const howItWorks = [
-  {
-    title: "Create Your Habits",
-    description: "Input your habit details and start tracking from day one",
-  },
-  {
-    title: "Visualize Your Progress",
-    description:
-      "Use the summary view for a quick overview and the calendar-like view for detailed history.",
-  },
-  {
-    title: "Celebrate Achievements",
-    description: "Reach milestones, earn badges and keep the motivation high.",
-  },
-];
-
-const HowItWorks = () => {
-  return (
-    <section id="how-it-works" className="py-10 md:py-20">
-      <h2 className="text-4xl font-bold text-white text-center mb-12 ">
-        How it works?
-      </h2>
-      <div className="select-none flex flex-col max-w-4xl mx-auto gap-10">
-        {howItWorks.map((item, index) => {
-          return (
-            <motion.div
-              key={item.title}
-              initial={{ opacity: 0 }}
-              whileInView={{ opacity: 1 }}
-              viewport={{ once: true, amount: "all" }}
-              className={cn(
-                `group  md:flex md:items-center md:gap-10  max-w-lg  `,
-                {
-                  "self-center": index === 1,
-                  "self-end": index === 2,
-                }
-              )}
-            >
-              <h3 className="group-hover:text-emerald-600 transition-colors text-5xl mb-4 text-zinc-400 font-bold  md:h-20 md:w-20 flex-shrink-0 md:grid md:place-content-center">
-                {index + 1}
-              </h3>
-              <div className="relative p-4 rounded-lg bg-gradient-to-tr from-white/5  border-[1px] border-white/10 before:h-full before:absolute  before:z-10 before:w-full before:bg-[url('/noise.svg')] before:content-[' '] before:block before:left-0 before:top-0 before:opacity-10">
-                <p className="text-white mb-1 text-lg font-medium ">
-                  {item.title}
-                </p>
-                <p className="text-zinc-400  font-medium">{item.description}</p>
-              </div>
-            </motion.div>
-          );
-        })}
-      </div>
-    </section>
-  );
-};
-
-const GetStarted = () => {
-  return (
-    <section
-      id="get-started"
-      className="text-center flex items-center flex-col py-10 md:py-20 "
-    >
-      <h2 className="text-4xl font-bold text-white mb-6">Get Started Today!</h2>
-      <p className="text-zinc-300 text-lg font-medium max-w-[50ch]">
-        Ready to embark on your habit transformation journey? Join DoguiHabits
-        now and take the first step towards a more disciplined, healthier, and
-        happier you.
-      </p>
-
-      <Link
-        className={
-          "bg-emerald-600 font-medium h-10 px-6 grid place-content-center rounded-md text-white mt-10"
-        }
-        to={"/login"}
-      >
-        Sign Up Now
-      </Link>
-    </section>
-  );
-};
-
-const footerLinks = [
-  {
-    name: "Github",
-    href: "https://github.com/GuidoCarda/doguihabits",
-    iconSrc: "/github.svg",
-  },
-  {
-    name: "LinkedIn",
-    href: "https://www.linkedin.com/in/guido-cardarelli/",
-    iconSrc: "/linkedin.svg",
-  },
-];
-
-const Footer = () => {
-  return (
-    <footer className="border-t-[1px] border-zinc-800 pt-10">
-      <div>
-        <div className="flex flex-col gap-10 items-center md:gap-0 md:flex-row md:justify-center">
-          <span className="font-bold text-xl text-zinc-100">Doguihabits</span>
-
-          <ul className="flex gap-10 md:ml-auto">
-            {footerLinks.map((item) => {
-              return (
-                <li key={item.name} className="">
-                  <a
-                    target="_black"
-                    rel="noopener"
-                    className="active:text-white font-medium hover:text-white transition-colors durations-100 text-zinc-400 flex gap-4 items-center"
-                    href={item.href}
-                  >
-                    <img src={item.iconSrc} className="h-8 w-8" />
-                    {item.name}
-                  </a>
-                </li>
-              );
-            })}
-          </ul>
-        </div>
-      </div>
-      <span className="font-medium text-zinc-400 mb-4 mt-16 md:mt-10 flex items-center justify-center">
-        Developed by
-        <a
-          className="mx-1 hover:text-emerald-500 transition-colors duration-300"
-          target="_blank"
-          rel="noopener"
-          href="https://github.com/guidocarda"
-        >
-          Guido Cardarelli
-        </a>{" "}
-        with
-        <HeartIcon className="ml-2 h-5 hover:scale-110 transition-transform w-5 text-emerald-500" />
-      </span>
-    </footer>
-  );
-};
-
-export default Landing;
+export default Features;
